Added joypad and fixed msvc compilation errors

+++ gb_lib/src/hram.cpp
@@ -6,7 +6,7 @@
 
 namespace gandalf {
     HRAM::HRAM() : Bus::AddressHandler("HRAM") {
-        data_.fill(std::rand());
+        data_.fill((byte)std::rand());
     }
 
     HRAM::~HRAM() = default;
